+++ autotorrent/tests/test_db.py
@@ -1,10 +1,12 @@
 from __future__ import unicode_literals
 
+import logging
 import os
 import shutil
 import tempfile
 
 from io import open
+from logging.handlers import BufferingHandler
 from unittest import TestCase
 
 from ..db import Database
@@ -17,6 +19,16 @@ def create_file(temp_folder, path, size):
     
     with open(path, 'w') as f:
         f.write(u'x' * size)
+
+class TestHandler(BufferingHandler):
+    def __init__(self):
+        BufferingHandler.__init__(self, 0)
+
+    def shouldFlush(self):
+        return False
+
+    def emit(self, record):
+        self.buffer.append(record.msg)
 
 class TestDatabase(TestCase):
     def setUp(self):
@@ -204,6 +216,20 @@ class TestDatabase(TestCase):
         self.db.clear_hash_size_table()
         self.db.build_hash_size_table()
         
-        self.assertEqual(self.db.find_hash_name('some-rls.mkv'),
-                         [os.path.join(self._temp_path, '3', 'Some-Release', 'Sample', 'some-rls.mkv'),
-                          os.path.join(self._temp_path, '3', 'Some-CD-Release', 'Sample', 'some-rls.mkv')])
+        self.assertEqual(sorted(self.db.find_hash_name('some-rls.mkv')),
+                         sorted([os.path.join(self._temp_path, '3', 'Some-Release', 'Sample', 'some-rls.mkv'),
+                          os.path.join(self._temp_path, '3', 'Some-CD-Release', 'Sample', 'some-rls.mkv')]))
+
+    def test_inaccessible_file(self):
+        h = TestHandler()
+        l = logging.getLogger('autotorrent.db')
+        l.addHandler(h)
+        
+        inaccessible_path = os.path.join(self._temp_path, '1', 'a')
+        os.chmod(inaccessible_path, 0000)
+        self.db.rebuild()
+        
+        self.assertTrue("Path %r is not accessible, skipping" % inaccessible_path in h.buffer)
+        
+        l.removeHandler(h)
+        h.close()