+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+Version 1.6.1 (11-05-2016)
+===========================================================
+
+*   Feature: Can now try and auto-detect local torrent clients
+*   Feature: Now possible to have multiple clients configured
+*   Bugfix: Fixed small bugs with Deluge client
+*   Bugfix: Dry run tried to connect to the client
+*   Bugfix: Encoding problems with transmission and Python 3
+
 Version 1.6.0 (30-04-2015)
 ===========================================================